requirements.txt
flask
groq
python-dotenv
edge-tts
gunicorn
